add support for halfrooms - closes #33

+++ src/Justimmo/Model/Mapper/V1/RealtyMapper.php
@@ -146,7 +146,7 @@ class RealtyMapper extends AbstractMapper
             ),
             'anzahl_zimmer'              => array(
                 'property' => 'roomCount',
-                'type'     => 'int',
+                'type'     => 'double',
             ),
             'anzahl_badezimmer'          => array(
                 'property' => 'bathroomCount',
